Organise classes into packages and create JUnit test classes
The classes used by BusynessManager are all in the main package.
Moreover, some of these classes do not contain JUnit tests.

Putting the classes into separate packages will help with keeping
the code organised and systematic. Implementing JUnit tests will
help to test the code for unexpected issues, thereby increasing
reliability.

Let's
* organise the different classes in BusynessManager into related
packages.
* create JUnit test classes for the classes in the main folder.
