+++ src/test/java/busynessmanager/managers/InventoryManagerTest.java
@@ -1,0 +1,4 @@
+package busynessmanager.managers;
+
+public class InventoryManagerTest {
+}
