+++ src/test/java/busynessmanager/managers/SalesManagerTest.java
@@ -1,0 +1,4 @@
+package busynessmanager.managers;
+
+public class SalesManagerTest {
+}
