+++ src/test/java/busynessmanager/managers/SearchManagerTest.java
@@ -1,0 +1,4 @@
+package busynessmanager.managers;
+
+public class SearchManagerTest {
+}
