+++ src/test/java/busynessmanager/product/ProductTest.java
@@ -1,0 +1,4 @@
+package busynessmanager.product;
+
+public class ProductTest {
+}
