+++ src/test/java/busynessmanager/revenue/RevenueCalculatorTest.java
@@ -1,0 +1,4 @@
+package busynessmanager.revenue;
+
+public class RevenueCalculatorTest {
+}
